+++ features/overlays/seaCreaturesCountAndTimeTracker.js
@@ -6,7 +6,7 @@ import { EntityArmorStand } from "../../constants/javaTypes";
 import { overlayCoordsData } from "../../data/overlayCoords";
 import { getWorldName, hasFishingRodInHotbar, isInHunterArmor, isInSkyblock } from "../../utils/playerState";
 import { CRIMSON_ISLE, CRYSTAL_HOLLOWS, HUB, KUUDRA } from "../../constants/areas";
-import { isInChatOrInventoryGui } from "../../utils/common";
+import { createButtonsDisplay, toggleButtonsDisplay } from "../../utils/overlays";
 
 const TIMER_THRESHOLD_IN_MINUTES = 5;
 
@@ -25,17 +25,7 @@ register("worldUnload", () => {
     resetTimer();
 });
 
-// DisplayLine is initialized once in order to avoid multiple method calls on click.
-let resetTrackerDisplay = new Display().hide();
-let resetTrackerDisplayLine = new DisplayLine(`${RED}[Click to reset]`).setShadow(true);
-resetTrackerDisplayLine.registerClicked((x, y, mouseButton, buttonState) => {
-    if (mouseButton === 0 && buttonState === false) { // When left mouse button is UP. 0 is left mouse button, false is UP, true is DOWN. 
-        resetTimer();
-    }
-});
-resetTrackerDisplayLine.registerHovered(() => resetTrackerDisplayLine.setText(`${RED}${UNDERLINE}[Click to reset]`).setShadow(true));
-resetTrackerDisplayLine.registerMouseLeave(() => resetTrackerDisplayLine.setText(`${RED}[Click to reset]`).setShadow(true));
-resetTrackerDisplay.addLine(resetTrackerDisplayLine);
+const buttonsDisplay = createButtonsDisplay(true, () => resetTimer(), false, null);
 
 function resetTimer() {
     startTime = null;
@@ -137,7 +127,7 @@ function renderCountOverlay() {
         !hasFishingRodInHotbar() ||
         allOverlaysGui.isOpen()
     ) {
-        resetTrackerDisplay.hide();
+        buttonsDisplay.hide();
         return;
     }
 
@@ -161,15 +151,7 @@ function renderCountOverlay() {
         .setScale(overlayCoordsData.seaCreaturesCountOverlay.scale);
     overlay.draw();
 
-    const shouldShowReset = isInChatOrInventoryGui();
-    if (shouldShowReset) {
-        resetTrackerDisplayLine.setScale(overlayCoordsData.seaCreaturesCountOverlay.scale - 0.2);
-        resetTrackerDisplay
-            .setRenderX(overlayCoordsData.seaCreaturesCountOverlay.x + overlay.getWidth() + 2)
-            .setRenderY(overlayCoordsData.seaCreaturesCountOverlay.y + 1).show();
-    } else {
-        resetTrackerDisplay.hide();
-    }
+    toggleButtonsDisplay(buttonsDisplay, overlay, overlayCoordsData.seaCreaturesCountOverlay);
 }
 
 function getSeaCreaturesCountThreshold() {